+++ stig_edit/__pycache__/__init__.cpython-311.pyc
